chore: Money 타입 Int -> Long 변경

+++ src/main/kotlin/model/Cashier.kt
@@ -1,7 +1,7 @@
 package model
 
 class Cashier {
-    fun toTicketQuantity(money: Money): Int = money.amount / LOTTO_PRICE
+    fun toTicketQuantity(money: Money): Int = (money.amount / LOTTO_PRICE).toInt()
 
     companion object {
         const val LOTTO_PRICE = 1000
